fix(css): Update Webpack, Babel configs, Fix Prod CSS-Modules, Wallaby
Fixed a long outstanding issue with CSS-Modules not working correctly in production builds. Turns out you need to explicitly set the context property on the base webpack config, in css-loader's options, and in the options for babel-plugin-react-css-modules (which has to be configured in webpack, not .babelrc). Yay for squashing huge bugs!

Updated babel config to remove redundant plugins provided by the stage-0 preset of babel 7. Removed all the redundant modules from package.json to keep from having to keep them all manually updated. Added stage-0 to the test env config, which solves for wallaby.js failing to start with an unexpected token export error, big thanks to Artem Govorov for bringing that mistake to my attention!

Fixed an error with #reducer and ModalReducer, where the reducer's name wasn't properly being used to setup and access state.